+++ R/data.R
@@ -14,3 +14,6 @@ NULL
 
 #' @rdname rcd_data_2021
 "rcd_2021_07_07"
+
+#' @rdname rcd_data_2021
+"rcd_2021_10_13"