+++ data_raw/rcd.R
@@ -3,12 +3,13 @@ system("git clone https://github.com/insightsengineering/random.cdisc.data.git")
 releases <- c(
   #"2021_03_22" = "v0.3.8",
   #"2021_05_05" = "v0.3.10",
-  "2021_07_07" = "v0.3.11"
+  #"2021_07_07" = "v0.3.11",
+  "2021_10_13" = "v0.3.12"
 )
 
-
-# https://stackoverflow.com/questions/5577221/how-can-i-load-an-object-into-a-variable-name-that-i-specify-from-an-r-data-file
-loadRData <- function(fileName){
+# https://stackoverflow.com/questions/5577221/
+#   how-can-i-load-an-object-into-a-variable-name-that-i-specify-from-an-r-data-file
+loadRData <- function(fileName) { # nolint
   #loads an RData file, and returns it
   load(fileName)
   get(ls()[ls() != "fileName"])
@@ -16,9 +17,7 @@ loadRData <- function(fileName){
 
 setwd("random.cdisc.data/")
 
-
 for (i in seq_along(releases)) {
-
   dt <- names(releases)[i]
   v <- releases[i]
 
@@ -36,7 +35,6 @@ for (i in seq_along(releases)) {
 
   cl <- call("save", as.name(nm), file = paste0("../data/", nm, ".RData"), compress = "bzip2")
   eval(cl)
-
 }
 
 setwd("..")