+++ NEWS.md
@@ -1,6 +1,7 @@
 # scda.2021 0.1.1.9007
 
 * Updated R version requirement to >= 3.6.
+* Added `rcd_2021_10_13` data to the package.
 
 # scda.2021 0.1.1